video_summerizer/requirements.txt
Flask
moviepy
openai-whisper
nltk
torch
werkzeug
imageio-ffmpeg